Initial Code for Fuzz Msg Generation

+++ src/ifl_msg.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif
 
-int IFL_CraftFuzzedMsg(IFL *ifl, uint8_t **out, uint16_t *out_len);
+int IFL_CraftFuzzedMsg(IFL *ifl, uint8_t **out, uint32_t *out_len);
 
 #ifdef __cplusplus
 }
